+++ Models/Ingredient.cs
@@ -3,7 +3,7 @@ namespace food_history_api.Models;
 public class Ingredient {
     public string Name {get; set;}
 
-    public int Quantity {get; set;}
+    public double Quantity {get; set;}
 
     public string? Uom {get; set;}
 
@@ -17,7 +17,7 @@ public class Ingredient {
         this.Uom = "ERROR";
     }
 
-    public Ingredient(string name, int quantity, string? uom, string? notes)
+    public Ingredient(string name, double quantity, string? uom, string? notes)
     {
         this.Name = name;
         this.Quantity = quantity;
